Restore reason to Storybook decorator
- Import indirection didn't seem to work out well
- Adding the decorator first resolves the provider issues
