+++ src/components/Blockquote.stories.tsx
@@ -1,4 +1,5 @@
-import '../storybook/register';
+import 'braid-design-system/reset';
+import 'loki/configure-react';
 
 import { List, Text } from 'braid-design-system';
 import React from 'react';
@@ -6,23 +7,20 @@ import { select } from 'sku/@storybook/addon-knobs';
 import { storiesOf } from 'sku/@storybook/react';
 
 import { withDecorator } from '../storybook/decorator';
-import { StorybookProvider } from '../storybook/provider';
 
 import { Blockquote } from './Blockquote';
 
 storiesOf('Blockquote', module)
+  .addDecorator(withDecorator)
   .add('Custom', () => {
     const size = select('size', ['standard', 'large'], 'standard');
 
     return (
-      <StorybookProvider>
-        <Blockquote size={size}>
-          <Text size={size}>This is a paragraph.</Text>
-          <List size={size}>
-            <Text size={size}>This is a bullet point.</Text>
-          </List>
-        </Blockquote>
-      </StorybookProvider>
+      <Blockquote size={size}>
+        <Text size={size}>This is a paragraph.</Text>
+        <List size={size}>
+          <Text size={size}>This is a bullet point.</Text>
+        </List>
+      </Blockquote>
     );
-  })
-  .addDecorator(withDecorator);
+  });
