+++ src/components/CopyableText.stories.tsx
@@ -1,36 +1,34 @@
-import '../storybook/register';
+import 'braid-design-system/reset';
+import 'loki/configure-react';
 
 import React from 'react';
 import { select, text } from 'sku/@storybook/addon-knobs';
 import { storiesOf } from 'sku/@storybook/react';
 
 import { withDecorator } from '../storybook/decorator';
-import { StorybookProvider } from '../storybook/provider';
 
 import { CopyableText } from './CopyableText';
 
 storiesOf('CopyableText', module)
+  .addDecorator(withDecorator)
   .add('Custom', () => (
-    <StorybookProvider>
-      <CopyableText
-        copiedLabel={select(
-          'copiedLabel',
-          { undefined, custom: 'Custom copied label' },
-          undefined,
-        )}
-        copyLabel={select(
-          'copyLabel',
-          { undefined, custom: 'Custom copy label' },
-          undefined,
-        )}
-        size={select(
-          'size',
-          ['xsmall', 'small', 'standard', 'large'],
-          'standard',
-        )}
-      >
-        {text('children', 'copy me')}
-      </CopyableText>
-    </StorybookProvider>
-  ))
-  .addDecorator(withDecorator);
+    <CopyableText
+      copiedLabel={select(
+        'copiedLabel',
+        { undefined, custom: 'Custom copied label' },
+        undefined,
+      )}
+      copyLabel={select(
+        'copyLabel',
+        { undefined, custom: 'Custom copy label' },
+        undefined,
+      )}
+      size={select(
+        'size',
+        ['xsmall', 'small', 'standard', 'large'],
+        'standard',
+      )}
+    >
+      {text('children', 'copy me')}
+    </CopyableText>
+  ));
