+++ src/components/InlineCode.stories.tsx
@@ -1,4 +1,5 @@
-import '../storybook/register';
+import 'braid-design-system/reset';
+import 'loki/configure-react';
 
 import { Text } from 'braid-design-system';
 import React from 'react';
@@ -6,20 +7,17 @@ import { select, text } from 'sku/@storybook/addon-knobs';
 import { storiesOf } from 'sku/@storybook/react';
 
 import { withDecorator } from '../storybook/decorator';
-import { StorybookProvider } from '../storybook/provider';
 
 import { InlineCode } from './InlineCode';
 
 storiesOf('InlineCode', module)
+  .addDecorator(withDecorator)
   .add('Custom', () => (
-    <StorybookProvider>
-      <Text>
-        Some text with{' '}
-        <InlineCode weight={select('weight', ['regular', 'weak'], 'regular')}>
-          {text('children', 'inline code')}
-        </InlineCode>
-        !
-      </Text>
-    </StorybookProvider>
-  ))
-  .addDecorator(withDecorator);
+    <Text>
+      Some text with{' '}
+      <InlineCode weight={select('weight', ['regular', 'weak'], 'regular')}>
+        {text('children', 'inline code')}
+      </InlineCode>
+      !
+    </Text>
+  ));
