+++ src/components/InternalLink.stories.tsx
@@ -1,4 +1,5 @@
-import '../storybook/register';
+import 'braid-design-system/reset';
+import 'loki/configure-react';
 
 import { Alert, Stack, Text } from 'braid-design-system';
 import React from 'react';
@@ -6,29 +7,23 @@ import { boolean, text } from 'sku/@storybook/addon-knobs';
 import { storiesOf } from 'sku/@storybook/react';
 
 import { withDecorator } from '../storybook/decorator';
-import { StorybookProvider } from '../storybook/provider';
 
 import { InternalLink } from './InternalLink';
 
 storiesOf('InternalLink', module)
+  .addDecorator(withDecorator)
   .add('Custom', () => (
-    <StorybookProvider>
-      <InternalLink
-        href={text('href', 'page#id')}
-        reset={boolean('reset', true)}
-      >
-        <Alert tone="caution">
-          <Stack space="gutter">
-            <Text>
-              {text('line1', 'InternalLink supports complex components.')}
-            </Text>
+    <InternalLink href={text('href', 'page#id')} reset={boolean('reset', true)}>
+      <Alert tone="caution">
+        <Stack space="gutter">
+          <Text>
+            {text('line1', 'InternalLink supports complex components.')}
+          </Text>
 
-            <Text size="small">
-              {text('line2', 'For example, this whole Alert is a link!')}
-            </Text>
-          </Stack>
-        </Alert>
-      </InternalLink>
-    </StorybookProvider>
-  ))
-  .addDecorator(withDecorator);
+          <Text size="small">
+            {text('line2', 'For example, this whole Alert is a link!')}
+          </Text>
+        </Stack>
+      </Alert>
+    </InternalLink>
+  ));
