+++ src/components/SmartTextLink.stories.tsx
@@ -1,4 +1,5 @@
-import '../storybook/register';
+import 'braid-design-system/reset';
+import 'loki/configure-react';
 
 import { Text } from 'braid-design-system';
 import React from 'react';
@@ -6,18 +7,15 @@ import { text } from 'sku/@storybook/addon-knobs';
 import { storiesOf } from 'sku/@storybook/react';
 
 import { withDecorator } from '../storybook/decorator';
-import { StorybookProvider } from '../storybook/provider';
 
 import { SmartTextLink } from './SmartTextLink';
 
 storiesOf('SmartTextLink', module)
+  .addDecorator(withDecorator)
   .add('Custom', () => (
-    <StorybookProvider>
-      <Text>
-        <SmartTextLink href={text('href', 'https://developer.seek.com/schema')}>
-          {text('children', 'SEEK Schema')}
-        </SmartTextLink>
-      </Text>
-    </StorybookProvider>
-  ))
-  .addDecorator(withDecorator);
+    <Text>
+      <SmartTextLink href={text('href', 'https://developer.seek.com/schema')}>
+        {text('children', 'SEEK Schema')}
+      </SmartTextLink>
+    </Text>
+  ));
