+++ src/components/Table.stories.tsx
@@ -1,4 +1,5 @@
-import '../storybook/register';
+import 'braid-design-system/reset';
+import 'loki/configure-react';
 
 import { Text } from 'braid-design-system';
 import React, { Fragment } from 'react';
@@ -6,71 +7,64 @@ import { select } from 'sku/@storybook/addon-knobs';
 import { storiesOf } from 'sku/@storybook/react';
 
 import { withDecorator } from '../storybook/decorator';
-import { StorybookProvider } from '../storybook/provider';
 
 import { Table } from './Table';
 import { TableRow } from './TableRow';
 
 storiesOf('Table', module)
+  .addDecorator(withDecorator)
   .add('Defaults', () => (
-    <StorybookProvider>
-      <Table header={['Column A', 'Column B']}>
-        <TableRow>
-          <Fragment>This is body cell A1.</Fragment>
-          <Fragment>B1</Fragment>
-        </TableRow>
+    <Table header={['Column A', 'Column B']}>
+      <TableRow>
+        <Fragment>This is body cell A1.</Fragment>
+        <Fragment>B1</Fragment>
+      </TableRow>
 
-        <TableRow>
-          <Fragment>A2</Fragment>
-          <Fragment>This is body cell B2.</Fragment>
-        </TableRow>
-      </Table>
-    </StorybookProvider>
+      <TableRow>
+        <Fragment>A2</Fragment>
+        <Fragment>This is body cell B2.</Fragment>
+      </TableRow>
+    </Table>
   ))
   .add('Stripe', () => (
-    <StorybookProvider>
-      <Table
-        align={['left', 'right']}
-        header={
-          <Fragment>
-            <Fragment>Column A</Fragment>
-            <Text weight="regular">Column B</Text>
-          </Fragment>
-        }
-        size={select('size', ['standard', 'large'], 'standard')}
-        type="stripe"
-        width={select('width', { undefined, full: 'full' }, undefined)}
-      >
-        <TableRow>
-          <Text>This is body cell A1.</Text>
-          <Text weight="strong">B1</Text>
-        </TableRow>
+    <Table
+      align={['left', 'right']}
+      header={
+        <Fragment>
+          <Fragment>Column A</Fragment>
+          <Text weight="regular">Column B</Text>
+        </Fragment>
+      }
+      size={select('size', ['standard', 'large'], 'standard')}
+      type="stripe"
+      width={select('width', { undefined, full: 'full' }, undefined)}
+    >
+      <TableRow>
+        <Text>This is body cell A1.</Text>
+        <Text weight="strong">B1</Text>
+      </TableRow>
 
-        <TableRow>
-          <Fragment>A2</Fragment>
-          <Fragment>This is body cell B2.</Fragment>
-        </TableRow>
-      </Table>
-    </StorybookProvider>
+      <TableRow>
+        <Fragment>A2</Fragment>
+        <Fragment>This is body cell B2.</Fragment>
+      </TableRow>
+    </Table>
   ))
   .add('Subtle', () => (
-    <StorybookProvider>
-      <Table
-        header={['Column A', 'Column B']}
-        size={select('size', ['standard', 'large'], 'standard')}
-        type="subtle"
-        width={select('width', { undefined, full: 'full' }, undefined)}
-      >
-        <TableRow>
-          <Fragment>This is body cell A1.</Fragment>
-          <Fragment>B1</Fragment>
-        </TableRow>
+    <Table
+      header={['Column A', 'Column B']}
+      size={select('size', ['standard', 'large'], 'standard')}
+      type="subtle"
+      width={select('width', { undefined, full: 'full' }, undefined)}
+    >
+      <TableRow>
+        <Fragment>This is body cell A1.</Fragment>
+        <Fragment>B1</Fragment>
+      </TableRow>
 
-        <TableRow>
-          <Fragment>A2</Fragment>
-          <Fragment>This is body cell B2.</Fragment>
-        </TableRow>
-      </Table>
-    </StorybookProvider>
-  ))
-  .addDecorator(withDecorator);
+      <TableRow>
+        <Fragment>A2</Fragment>
+        <Fragment>This is body cell B2.</Fragment>
+      </TableRow>
+    </Table>
+  ));
