+++ src/components/WrapperRenderer.stories.tsx
@@ -1,4 +1,5 @@
-import '../storybook/register';
+import 'braid-design-system/reset';
+import 'loki/configure-react';
 
 import { Text } from 'braid-design-system';
 import React from 'react';
@@ -6,20 +7,15 @@ import { storiesOf } from 'sku/@storybook/react';
 
 import { withDecorator } from '../storybook/decorator';
 import Wrapper from '../storybook/markdown/wrapper.mdx';
-import { StorybookProvider } from '../storybook/provider';
 
 import { WrapperRenderer } from './WrapperRenderer';
 
 storiesOf('WrapperRenderer', module)
+  .addDecorator(withDecorator)
   .add('Example', () => (
-    <StorybookProvider>
-      <WrapperRenderer document={Wrapper}>
-        {({ children }) => (
-          <Text>
-            {React.Children.toArray(children).length} top-level node(s)
-          </Text>
-        )}
-      </WrapperRenderer>
-    </StorybookProvider>
-  ))
-  .addDecorator(withDecorator);
+    <WrapperRenderer document={Wrapper}>
+      {({ children }) => (
+        <Text>{React.Children.toArray(children).length} top-level node(s)</Text>
+      )}
+    </WrapperRenderer>
+  ));
