+++ src/storybook/decorator.tsx
@@ -1,16 +1,42 @@
-import React from 'react';
+import { select, text } from '@storybook/addon-knobs';
+import { BraidLoadableProvider, Card, ContentBlock } from 'braid-design-system';
+import React, { ReactNode } from 'react';
 import { Helmet } from 'react-helmet';
 import { BrowserRouter } from 'react-router-dom';
-import { select, text } from 'sku/@storybook/addon-knobs';
 import { addDecorator } from 'sku/@storybook/react';
 
-import { MdxProvider } from '..';
+import { MdxProvider, ScoobieLink } from '..';
 import { robotoHref, robotoMonoHref } from '../../typography';
 import { DEFAULT_SIZE, SIZES } from '../private/size';
 
-type DecoratorFunction = Parameters<typeof addDecorator>[0];
+interface Props {
+  children: ReactNode;
+}
 
-export const withDecorator: DecoratorFunction = (story) => (
+const BraidStorybookProvider = ({ children }: Props) => (
+  <BraidLoadableProvider
+    linkComponent={ScoobieLink}
+    themeName={select(
+      'BraidLoadableProvider.themeName',
+      [
+        'apac',
+        'catho',
+        'docs',
+        'jobsDb',
+        'jobStreet',
+        'occ',
+        'seekAnz',
+        'seekBusiness',
+        'wireframe',
+      ],
+      'apac',
+    )}
+  >
+    {children}
+  </BraidLoadableProvider>
+);
+
+const MdxStorybookProvider = ({ children }: Props) => (
   <MdxProvider
     graphqlPlayground={
       text(
@@ -20,13 +46,25 @@ export const withDecorator: DecoratorFunction = (story) => (
     }
     size={select('MdxProvider.size', SIZES, DEFAULT_SIZE)}
   >
-    <BrowserRouter>
-      <Helmet>
-        <link href={robotoHref} rel="stylesheet" />
-        <link href={robotoMonoHref} rel="stylesheet" />
-      </Helmet>
-
-      {story()}
-    </BrowserRouter>
+    {children}
   </MdxProvider>
+);
+
+type DecoratorFunction = Parameters<typeof addDecorator>[0];
+
+export const withDecorator: DecoratorFunction = (story) => (
+  <BrowserRouter>
+    <Helmet>
+      <link href={robotoHref} rel="stylesheet" />
+      <link href={robotoMonoHref} rel="stylesheet" />
+    </Helmet>
+
+    <BraidStorybookProvider>
+      <MdxStorybookProvider>
+        <ContentBlock>
+          <Card>{story()}</Card>
+        </ContentBlock>
+      </MdxStorybookProvider>
+    </BraidStorybookProvider>
+  </BrowserRouter>
 );
