--- src/storybook/register.ts
@@ -1,2 +1,0 @@
-import 'braid-design-system/reset';
-import 'loki/configure-react';
